04-prev-unknown-vulns/building/requirements.txt
GitPython
docker
